crashing bug + clean up + misc updates
- fixed crashing bug around buffer handing
- swapped moisture and voltage so that wifi won't turn off voltage
- got rid of more String instances
- now saves hub address and ports in flash
- saves if it rebooted because it couldn't connect
- updated some logging to be more clear  (e.g. converted epoch to actual time)

+++ SmartThings_Edge/Arduino/WeatherStation/BME280Handler.ino
@@ -2,7 +2,7 @@
 #include "EnvironmentCalculations.h"
 
 extern Logger logger;
-extern int currentTime();
+extern unsigned long currentTime();
 
 void BME280Handler::init()
 {
@@ -39,7 +39,7 @@ float BME280Handler::getHumidity()
   return humidity;
 }
 
-long BME280Handler::getReadingTime()
+unsigned long BME280Handler::getReadingTime()
 {
   return readingTime;
 }
